+++ core-ng/src/main/java/core/framework/api/http/HTTPHeaders.java
@@ -9,4 +9,5 @@ public final class HTTPHeaders {
     public static final String LOCATION = "Location";
     public static final String IF_MODIFIED_SINCE = "If-Modified-Since";
     public static final String LAST_MODIFIED = "Last-Modified";
+    public static final String USER_AGENT = "User-Agent";
 }