+++ core-ng/src/main/java/core/framework/impl/queue/SNSMessagePublisher.java
@@ -42,8 +42,8 @@ public class SNSMessagePublisher<T> implements MessagePublisher<T> {
                     new MessageAttributeValue().withDataType("String").withStringValue("sns"))
                 .addMessageAttributesEntry(SQSMessageListener.MESSAGE_ATTR_TYPE,
                     new MessageAttributeValue().withDataType("String").withStringValue(messageType))
-                .addMessageAttributesEntry(SQSMessageListener.MESSAGE_ATTR_SENDER,
-                    new MessageAttributeValue().withDataType("String").withStringValue(Network.localHostName()));
+                .addMessageAttributesEntry(SQSMessageListener.MESSAGE_ATTR_CLIENT_IP,
+                    new MessageAttributeValue().withDataType("String").withStringValue(Network.localHostAddress()));
 
             linkContext(request);
 
@@ -64,7 +64,7 @@ public class SNSMessagePublisher<T> implements MessagePublisher<T> {
         ActionLog actionLog = logManager.currentActionLog();
         if (actionLog == null) return;
 
-        request.addMessageAttributesEntry(SQSMessageListener.MESSAGE_ATTR_REQUEST_ID,
+        request.addMessageAttributesEntry(SQSMessageListener.MESSAGE_ATTR_REF_ID,
             new MessageAttributeValue().withDataType("String").withStringValue(actionLog.refId()));
 
         if (actionLog.trace) {