+++ core-ng/src/main/java/core/framework/impl/queue/SQSMessageListener.java
@@ -36,8 +36,9 @@ public class SQSMessageListener implements Runnable, MessageHandlerConfig {
     // to keep backward compatible with core 2.x, SQS/SNS message will be replaced by RabbitMQ
     static final String MESSAGE_ATTR_PUBLISHER = "event_publisher";
     static final String MESSAGE_ATTR_TYPE = "event_type";
-    static final String MESSAGE_ATTR_SENDER = "event_sender";
-    static final String MESSAGE_ATTR_REQUEST_ID = "request_id";
+
+    static final String MESSAGE_ATTR_CLIENT_IP = "clientIP";
+    static final String MESSAGE_ATTR_REF_ID = "ref-id";
     static final String MESSAGE_ATTR_TRACE = "trace";
 
     private final ExecutorService listenerExecutor = Executors.newSingleThreadExecutor();
@@ -129,7 +130,7 @@ public class SQSMessageListener implements Runnable, MessageHandlerConfig {
 
     private <T> void process(Message sqsMessage) throws Exception {
         ActionLog actionLog = logManager.currentActionLog();
-        actionLog.action = "queue/" + queueName;
+        actionLog.action("queue/" + queueName);
 
         logger.debug("queueURL={}", queueURL);
         logger.debug("message={}", sqsMessage);
@@ -174,30 +175,28 @@ public class SQSMessageListener implements Runnable, MessageHandlerConfig {
     }
 
     private void linkSNSContext(ActionLog actionLog, SNSMessage snsMessage) {
-        if (snsMessage.attributes.requestId != null) {
-            actionLog.refId(snsMessage.attributes.requestId.value);
+        if (snsMessage.attributes.refId != null) {
+            actionLog.refId(snsMessage.attributes.refId.value);
         }
         if (snsMessage.attributes.trace != null && "true".equals(snsMessage.attributes.trace.value)) {
-            logger.warn("trace log is triggered for current sqs message, logId={}", actionLog.id);
-            actionLog.trace = true;
+            actionLog.triggerTraceLog();
         }
-        if (snsMessage.attributes.eventSender != null) {
-            actionLog.putContext("sender", snsMessage.attributes.eventSender.value);
+        if (snsMessage.attributes.clientIP != null) {
+            actionLog.putContext("clientIP", snsMessage.attributes.clientIP.value);
         }
     }
 
     private void linkSQSContext(ActionLog actionLog, Map<String, MessageAttributeValue> attributes) {
-        MessageAttributeValue requestId = attributes.get(MESSAGE_ATTR_REQUEST_ID);
-        if (requestId != null) actionLog.refId(requestId.getStringValue());
+        MessageAttributeValue refId = attributes.get(MESSAGE_ATTR_REF_ID);
+        if (refId != null) actionLog.refId(refId.getStringValue());
 
         MessageAttributeValue trace = attributes.get(MESSAGE_ATTR_TRACE);
         if (trace != null && "true".equals(trace.getStringValue())) {
-            logger.warn("trace log is triggered for current sqs message, logId={}", actionLog.id);
-            actionLog.trace = true;
+            actionLog.triggerTraceLog();
         }
-        MessageAttributeValue sender = attributes.get(MESSAGE_ATTR_SENDER);
-        if (sender != null) {
-            actionLog.putContext("sender", sender.getStringValue());
+        MessageAttributeValue clientIP = attributes.get(MESSAGE_ATTR_CLIENT_IP);
+        if (clientIP != null) {
+            actionLog.putContext("clientIP", clientIP.getStringValue());
         }
     }
 
@@ -207,8 +206,8 @@ public class SQSMessageListener implements Runnable, MessageHandlerConfig {
             .withMaxNumberOfMessages(10)
             .withMessageAttributeNames(MESSAGE_ATTR_PUBLISHER,
                 MESSAGE_ATTR_TYPE,
-                MESSAGE_ATTR_SENDER,
-                MESSAGE_ATTR_REQUEST_ID,
+                MESSAGE_ATTR_CLIENT_IP,
+                MESSAGE_ATTR_REF_ID,
                 MESSAGE_ATTR_TRACE));
 
         List<Message> messages = result.getMessages();